cli/requirements.txt
web3
py-solc